// frontend/src/types/student/profile.ts
export const socialLinkSchema = z.object({
    platform: z.string().min(1, "Platform is required"),
    url: z.url('Invalid URL').or(z.literal('')),
});
// ...
export type SocialLinkInput = z.infer<typeof socialLinkSchema>;
export type UpdateSocialLinkInput = Partial<SocialLinkInput>;
export type SocialLink = SocialLinkInput & { id: number; profileId: number };